refac: move test cases to test file
